New delta scatter plot with defeat annotations
This scatter plot shows the barometer deltas, so if the pressure dropped in the month before the battle, the line for that battle shows as blue. If the pressure increased, the line shows as orange.

Napoleon's rare defeats are shown as annotations. Currently they don't seem to be aligning properly with the x-ticks (battle names) - TODO fix.